microservices_iskrasvet/auth_service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyjwt==2.3.0
werkzeug==2.0.1 